WIP: Remove managed fields when comparing plan
This change was made to try to resolve upstream GH
hashicorp/terraform-provider-kubernetes-alpha#208

I don't know if this is correct code, so if you merge this you're
on your own. I mean, even more so because you're using alpha software!

There's definitely at least one problem, that I don't know how to
resolve yet.

If you have a resource on the server that has an object and your
kubernetes_manifest resource does not reference that object, you'll
get a "wrong final value type: incorrect object attributes" error.
Similarly, if you have a resource on the server that does not have
an object that you've specified in your TF file, you may see a
different error.

In either situation, add an object with a dummy value to your TF resource.

For example, say you have this resource on the server:

  apiVersion: v1
  kind: Whatever
  metadata:
    name: whatever
    annotations:
      foo: bar

or the resource on the server doesn't have any annotations. On
apply (not plan) you'll get one of the two errors mentioned above.

The workaround? Add a dummy hacky annotation.

Sorry for the verbosity, these notes are mostly for me in case I
forget why I did the thing I did.